# Faster Log Subscriber Check