@schneems 2017-11-02
Can't merge. People are using that constant to modify the behavior of their apps https://github.com/search?l=Ruby&q=IGNORE_PAYLOAD_NAMES&type=Code&utf8=

thanks  @matthewd  @pixeltrix