@jeremy 2017-11-02
Good reminder to `private_constant :INTERNAL_THINGS` 🤫