@rafaelfranca 2017-11-02
An that we are lacking a public API for that if it is so common.